ai_human_game/requirements.txt
numpy
python-dotenv
openai
pyfiglet 